requirements.txt
asgiref==3.2.10
attrs==19.3.0
awsebcli==3.18.1
Babel==2.9.0
bcrypt==3.1.7
blessed==1.17.8
botocore==1.15.49
cached-property==1.5.1
cart==1.1.8
cement==2.8.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
Django==3.0.8
django-admin-sortable2==0.7.6
django-angular==2.3
django-appconf==1.0.4
django-bootstrap3==14.2.0
django-cart==1.0.4
django-classy-tags==1.0.0
django-cms==3.7.3
django-cors-headers==3.4.0
django-cors-middleware==1.5.0
django-crispy-forms==1.9.2
django-csvimport==2.16
django-datetime-widget==0.9.3
django-easy-maps==1.1.1
django-entangled==0.3
django-filer==1.7.1
django-filter==2.3.0
django-formtools==2.2
django-fsm==2.7.0
django-fsm-admin==1.2.4
django-instagram==0.3.2
django-ipware==3.0.0
django-jet==1.0.8
django-js-asset==1.2.2
django-map-widgets==0.2.0
django-mathfilters==1.0.0
django-mptt==0.11.0
django-oauth-toolkit==1.3.2
django-paypal==1.0.0
django-pickling==1.0
django-polymorphic==2.1.2
django-post-office==3.4.1
django-rest-auth==0.9.5
django-searchable-select==1.5.0
django-sekizai==1.1.0
django-select2==7.4.2
django-shop==1.2.2
django-shopify-auth==0.9.1
django-shopping-cart==0.1
django-treebeard==4.3.1
django-vote==2.2.0
django-widget-tweaks==1.4.8
djangocms-admin-style==1.5.0
djangocms-cascade==1.3.2
djangocms-text-ckeditor==3.9.1
djangorestframework==3.11.0
djangorestframework-gis==0.15
docker==4.2.2
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drfdocs @ git+https://github.com/mago960806/django-rest-framework-docs.git@7f8dbff17ebc5cdaa6c1741ee62f982cf9eed9b1
easy-thumbnails==2.7
emoji==0.5.4
FinQuant==0.2.2
FundamentalAnalysis==0.2.6
future==0.16.0
geographiclib==1.50
geopy==2.0.0
html5lib==1.1
idna==2.7
importlib-metadata==1.7.0
inflection==0.5.1
iniconfig==1.1.1
jmespath==0.10.0
joblib==1.0.0
json-table-schema==0.2.1
jsonfield==3.1.0
jsonschema==3.2.0
kiwisolver==1.3.1
<<<<<<< HEAD
lightgbm 
=======
lightgbm @ file:///home/greats/Downloads/lightgbm-3.1.1-py2.py3-none-manylinux1_x86_64.whl
>>>>>>> fe2c9bd2d5d9d693e3b134dfde94bb3dc2d99c4d
lxml==4.5.2
Markdown==3.3.3
matplotlib==3.3.3
messytables==0.15.2
more-itertools==8.6.0
multitasking==0.0.9
numpy==1.19.5
oauthlib==3.1.0
packaging==20.8
pandas==1.2.0
paramiko==2.7.1
pathspec==0.5.9
paynow==1.0.4
Pillow==8.1.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.10.0
pyactiveresource==2.2.1
pycparser==2.20
pycryptodome==3.9.8
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.2.1
python-dateutil==2.8.0
python-dotenv==0.15.0
python-http-client==3.2.7
python-magic==0.4.18
pytz==2020.1
PyYAML==5.3.1
Quandl==3.5.3
requests==2.20.1
<<<<<<< HEAD
scikit-learn 
scipy 
=======
scikit-learn @ file:///home/greats/Downloads/scikit_learn-0.24.1-cp38-cp38-manylinux2010_x86_64.whl
scipy @ file:///home/greats/Downloads/scipy-1.6.0-cp38-cp38-manylinux1_x86_64.whl
>>>>>>> fe2c9bd2d5d9d693e3b134dfde94bb3dc2d99c4d
seaborn==0.11.1
semantic-version==2.5.0
sendgrid==6.4.3
ShopifyAPI==8.0.0
simplejson==3.17.2
six==1.11.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
stripe==2.48.0
termcolor==1.1.0
texttable==1.6.2
threadpoolctl==2.1.0
toml==0.10.2
twilio==6.44.0
Unidecode==1.1.1
urllib3==1.24.3
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
xlrd==2.0.1
yfinance==0.1.55
zipp==3.1.0
